Move actions around.
