+++ crates/node/platform/src/actions/install_deps.rs
@@ -1,73 +1,16 @@
-use moon_config::{NodeConfig, NodePackageManager, NodeVersionManager};
-use moon_error::MoonError;
+use moon_config::NodePackageManager;
 use moon_lang::has_vendor_installed_dependencies;
 use moon_logger::{debug, warn};
-use moon_node_lang::{PackageJson, NODE, NODENV, NPM, NVM};
+use moon_node_lang::NODE;
 use moon_node_tool::NodeTool;
 use moon_terminal::{print_checkpoint, Checkpoint};
 use moon_tool::ToolError;
 use moon_utils::{is_ci, is_test_env};
-use starbase_styles::color;
-use starbase_utils::fs;
 use std::path::Path;
 
 const LOG_TARGET: &str = "moon:node-platform:install-deps";
 
-/// Add `packageManager` to `package.json`.
-fn add_package_manager(node_config: &NodeConfig, package_json: &mut PackageJson) -> bool {
-    let manager_version = match node_config.package_manager {
-        NodePackageManager::Npm => node_config.npm.version.as_ref().map(|v| format!("npm@{v}")),
-        NodePackageManager::Pnpm => node_config.pnpm.as_ref().map(|cfg| {
-            cfg.version
-                .as_ref()
-                .map(|v| format!("pnpm@{v}"))
-                .unwrap_or_default()
-        }),
-        NodePackageManager::Yarn => node_config.yarn.as_ref().map(|cfg| {
-            cfg.version
-                .as_ref()
-                .map(|v| format!("yarn@{v}"))
-                .unwrap_or_default()
-        }),
-    };
-
-    if let Some(version) = manager_version {
-        if package_json.set_package_manager(&version) {
-            debug!(
-                target: LOG_TARGET,
-                "Adding package manager version to {}",
-                color::file(NPM.manifest)
-            );
-
-            return true;
-        }
-    }
-
-    false
-}
-
-/// Add `engines` constraint to `package.json`.
-fn add_engines_constraint(node_config: &NodeConfig, package_json: &mut PackageJson) -> bool {
-    if let Some(node_version) = &node_config.version {
-        if node_config.add_engines_constraint && package_json.add_engine("node", node_version) {
-            debug!(
-                target: LOG_TARGET,
-                "Adding engines version constraint to {}",
-                color::file(NPM.manifest)
-            );
-
-            return true;
-        }
-    }
-
-    false
-}
-
-pub async fn install_deps(
-    node: &NodeTool,
-    working_dir: &Path,
-    workspace_root: &Path,
-) -> Result<(), ToolError> {
+pub async fn install_deps(node: &NodeTool, working_dir: &Path) -> Result<(), ToolError> {
     // When in CI, we can avoid installing dependencies because
     // we can assume they've already been installed before moon runs!
     if is_ci() && has_vendor_installed_dependencies(working_dir, &NODE) {
@@ -77,35 +20,6 @@ pub async fn install_deps(
         );
 
         return Ok(());
-    }
-
-    // Sync values to `package.json`
-    if working_dir == workspace_root {
-        PackageJson::sync(working_dir, |package_json| {
-            let added_manager = add_package_manager(&node.config, package_json);
-            let added_constraint = add_engines_constraint(&node.config, package_json);
-
-            Ok(added_manager || added_constraint)
-        })?;
-    }
-
-    // Create nvm/nodenv version file
-    if let Some(version_manager) = &node.config.sync_version_manager_config {
-        if let Some(node_version) = &node.config.version {
-            let rc_name = match version_manager {
-                NodeVersionManager::Nodenv => NODENV.version_file.to_string(),
-                NodeVersionManager::Nvm => NVM.version_file.to_string(),
-            };
-            let rc_path = working_dir.join(rc_name);
-
-            fs::write_file(&rc_path, node_version).map_err(MoonError::StarFs)?;
-
-            debug!(
-                target: LOG_TARGET,
-                "Syncing Node.js version to {}",
-                color::path(&rc_path)
-            );
-        }
     }
 
     let package_manager = node.get_package_manager();
