+++ crates/node/platform/src/actions/mod.rs
@@ -1,7 +1,9 @@
 mod install_deps;
 mod run_target;
+mod setup_tool;
 mod sync_project;
 
 pub use install_deps::*;
 pub use run_target::*;
+pub use setup_tool::*;
 pub use sync_project::*;
